+++ docs/CHANGELOG.rst
@@ -28,7 +28,8 @@ Added
 - Improve log messages in executor and workload connectors
 - Add ``process_memory`` and ``process_cores`` fields to ``Data`` model and
   ``DataSerializer``
-- Support lookup expressions (``lt``, ``lte``, ``gt``, ``gte``) in ES viewsets
+- Support lookup expressions (``lt``, ``lte``, ``gt``, ``gte``, ``in``) in ES
+  viewsets
 - Support for easier dynamic composition of type extensions
 - Add ``elastic_mapping`` management command